modified public sidebar
