+++ src/app/components/pages/landing/page.jsx
@@ -1,19 +1,21 @@
 'use client'
 import { useSession } from 'next-auth/react'
 
+
 // pages
 import LandingSideBarPublic from '../../sidebar/landingSideBar/landingSideBarPublic'
 import LandingSideBarLogged from '../../sidebar/landingSideBar/landingSideBarLogged'
 import FeaturedStreamers from './featuredStreamers/featuredStremers'
 
-const LandingRender = ({streamers}) => {
+const LandingRender = () => {
 
   let session = useSession()
 
+
   return (
     <div className='flex flex-row flex-nowrap h-full'>
-      { session.status === 'authenticated' ? <LandingSideBarLogged streamers={streamers} session={session}/> : <LandingSideBarPublic streamers={streamers} /> }
-      <FeaturedStreamers streamers={streamers} />
+      { session.status === 'authenticated' ? <LandingSideBarLogged session={session}/> : <LandingSideBarPublic /> }
+      <FeaturedStreamers />
     </div>
   );
 }
